requirements.txt
Flask==2.3.3
redis==5.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
validators==0.22.0
